# backend_openapi_v3.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "NECO Photo Album API",
                 "description":  "FastAPI backend for NECO photo album generation",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/v1/albums/generate":  {
                                                  "post":  {
                                                               "tags":  [
                                                                            "albums"
                                                                        ],
                                                               "summary":  "Generate Album",
                                                               "operationId":  "generate_album_api_v1_albums_generate_post",
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/GenerateAlbumRequest"
                                                                                                                                       }
                                                                                                                        }
                                                                                               },
                                                                                   "required":  true
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "Successful Response",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {

                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "422":  {
                                                                                             "description":  "Validation Error",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         }
                                                                             }
                                                           }
                                              },
                  "/api/v1/albums/{album_id}/download":  {
                                                             "get":  {
                                                                         "tags":  [
                                                                                      "albums"
                                                                                  ],
                                                                         "summary":  "Download Album",
                                                                         "operationId":  "download_album_api_v1_albums__album_id__download_get",
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "album_id",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "string",
                                                                                                               "title":  "Album Id"
                                                                                                           }
                                                                                            }
                                                                                        ],
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "Successful Response",
                                                                                                       "content":  {
                                                                                                                       "application/json":  {
                                                                                                                                                "schema":  {

                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                   }
                                                                                                   },
                                                                                           "422":  {
                                                                                                       "description":  "Validation Error",
                                                                                                       "content":  {
                                                                                                                       "application/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                   }
                                                                                                   }
                                                                                       }
                                                                     }
                                                         },
                  "/api/v1/albums/{album_id}":  {
                                                    "delete":  {
                                                                   "tags":  [
                                                                                "albums"
                                                                            ],
                                                                   "summary":  "Delete Album",
                                                                   "operationId":  "delete_album_api_v1_albums__album_id__delete",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "album_id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "string",
                                                                                                         "title":  "Album Id"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Successful Response",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {

                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             },
                                                                                     "422":  {
                                                                                                 "description":  "Validation Error",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                },
                  "/api/v1/albums/generate-to-disk":  {
                                                          "post":  {
                                                                       "tags":  [
                                                                                    "albums"
                                                                                ],
                                                                       "summary":  "Generate Albums To Disk",
                                                                       "operationId":  "generate_albums_to_disk_api_v1_albums_generate_to_disk_post",
                                                                       "requestBody":  {
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/AlbumGenerationToDiskRequest"
                                                                                                                                               }
                                                                                                                                }
                                                                                                       },
                                                                                           "required":  true
                                                                                       },
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "Successful Response",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {

                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 },
                                                                                         "422":  {
                                                                                                     "description":  "Validation Error",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                      },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "GenerateAlbumRequest":  {
                                                                    "properties":  {
                                                                                       "school_id":  {
                                                                                                         "anyOf":  [
                                                                                                                       {
                                                                                                                           "type":  "string"
                                                                                                                       },
                                                                                                                       {
                                                                                                                           "type":  "null"
                                                                                                                       }
                                                                                                                   ],
                                                                                                         "title":  "School Id"
                                                                                                     },
                                                                                       "state_code":  {
                                                                                                          "anyOf":  [
                                                                                                                        {
                                                                                                                            "type":  "string"
                                                                                                                        },
                                                                                                                        {
                                                                                                                            "type":  "null"
                                                                                                                        }
                                                                                                                    ],
                                                                                                          "title":  "State Code"
                                                                                                      },
                                                                                       "batch":  {
                                                                                                     "anyOf":  [
                                                                                                                   {
                                                                                                                       "type":  "string"
                                                                                                                   },
                                                                                                                   {
                                                                                                                       "type":  "null"
                                                                                                                   }
                                                                                                               ],
                                                                                                     "title":  "Batch"
                                                                                                 },
                                                                                       "format":  {
                                                                                                      "type":  "string",
                                                                                                      "title":  "Format",
                                                                                                      "default":  "pdf"
                                                                                                  },
                                                                                       "layout":  {
                                                                                                      "type":  "string",
                                                                                                      "title":  "Layout",
                                                                                                      "default":  "grid_3x4"
                                                                                                  },
                                                                                       "include_missing_photos":  {
                                                                                                                      "type":  "boolean",
                                                                                                                      "title":  "Include Missing Photos",
                                                                                                                      "default":  false
                                                                                                                  }
                                                                                   },
                                                                    "type":  "object",
                                                                    "title":  "GenerateAlbumRequest"
                                                                },
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                   }
                   }
}